+++ src/main/java/racing_car_with_winner/model/CarFactory.java
@@ -4,6 +4,9 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import racing_car_with_winner.model.car.Car;
+import racing_car_with_winner.model.car.CarName;
+
 public class CarFactory {
     private final CarNameInput carNameInput;
