+++ src/main/java/racing_car_with_winner/model/car/Cars.java
@@ -1,9 +1,11 @@
-package racing_car_with_winner.model;
+package racing_car_with_winner.model.car;
 
 import java.util.Comparator;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
+
+import racing_car_with_winner.model.MovableStrategy;
 
 public class Cars {
     private final List<Car> cars;
