+++ src/test/java/racing_car_with_winner/model/CarTest.java
@@ -3,6 +3,9 @@ package racing_car_with_winner.model;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 
+import racing_car_with_winner.model.car.Car;
+import racing_car_with_winner.model.car.CarName;
+
 public class CarTest {
 
     @Test
