+++ src/test/java/racing_car_with_winner/model/MoveResultTest.java
@@ -1,9 +1,9 @@
 package racing_car_with_winner.model;
 
-import static org.junit.jupiter.api.Assertions.*;
-
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
+
+import racing_car_with_winner.model.car.MoveResult;
 
 class MoveResultTest {
     @Test
